+++ metautil.d.ts
@@ -46,6 +46,7 @@ export function duration(s: string | number): number;
 export function bytesToSize(bytes: number): string;
 export function sizeToBytes(size: string): number;
 export function namespaceByPath(namespace: object, path: string): object | null;
+export function deflatObject(source: object, fieldName: string): object;
 
 type Every = {
   month: number;